DocumentationOverview.jl

Summary

Name	Summary

API	An object passed to include callback of table.

find	List APIs in module.

table	Show the list of APIs as a table.

table_md	A shorthand for Markdown.MD(DocumentationOverview.table(args...; options...)).

Note

The above table is generated using

```@eval
using DocumentationOverview
DocumentationOverview.table_md(
    DocumentationOverview;
    signature = :name,
    apicolumn = "Name",
)
```

For more examples, see Gallery.

Documentation

DocumentationOverview.table — Function

DocumentationOverview.table(module::Module; [apicolumn], [include]) -> table
DocumentationOverview.table(apis; [apicolumn]) -> table

Show the list of APIs as a table.

The table object returned from this function supports show method with text/plain, text/markdown, and text/html MIME types.

To preprocess the list of APIs, get a vector of APIs from find, process it, and pass it to DocumentationOverview.table as an iterable of APIs.

See the Gallery for example outputs.

Extended help

Keyword Arguments

apicolumn::AbstractString = "**API**": The title for the API column.

signature: following values are supported:

nothing (default): use the default signature.

:name: use api.name; i.e., A.B.C.f(x, y) becomes f.

:strip_namespace: strip the namespace part; i.e., A.B.C.f(x, y) becomes f(x, y).

a callabel object: it must map an API to a String which is used as a signature.

Examples

julia> using DocumentationOverview

julia> DocumentationOverview.table(DocumentationOverview)
... a table of API printed ...

The list of APIs retrieved using PublicAPI.jl API can also be used with DocumentationOverview.table:

julia> using DocumentationOverview, PublicAPI

julia> table = DocumentationOverview.table(PublicAPI.of(DocumentationOverview));

DocumentationOverview.table_md — Function

DocumentationOverview.table_md(args...; options...) -> md::Markdown.MD

A shorthand for Markdown.MD(DocumentationOverview.table(args...; options...)).

DocumentationOverview.find — Function

DocumentationOverview.find(module::Module; [include]) -> apis::Vector{<:API}

List APIs in module.

Extended help

Keyword Arguments

include: a callable of form include(api::API) -> should_include::Bool. The apis are included in the output if and only if this function returns true. The default function is eauivalent to api -> api.hasdoc.

Examples

julia> using DocumentationOverview

julia> for api in DocumentationOverview.find(DocumentationOverview)
           @show api.name
       end
api.name = :API
api.name = :find
api.name = :table

DocumentationOverview.API — Type

DocumentationOverview.API

An object passed to include callback of table.

API has the following properties settable using Accessors.@set:

module::Module: the module in which the API is discovered.

name::Symbol: the name of the API.

exported::Bool: true iff the API is exported from api.module.

signature::String: "signature" of the API; if api.hasdoc (see below), it defaults to the first line of the docstring; otherwise, it defaults to the fully qualified name of the API.

It also has the following derived properties that are not settable with Accessors.@set:

hasdoc::Bool: indicate if the API has a docstring.

doc::Union{Markdown.MD,Nothing}: docstring of the API.